# docs/openapi3.json
{
  "components": {
...
    "schemas": {
...
      "github_com_ialekseychuk_my-place_internal_dto.CreateStaffRequest": {
        "properties": {
          "description": {
            "maxLength": 500,
            "type": "string"
          },
          "first_name": {
            "maxLength": 50,
            "minLength": 2,
            "type": "string"
          },
          "gender": {
            "enum": [
              "male",
              "female",
              "other",
              "prefer_not_to_say"
            ],
            "type": "string"
          },
          "last_name": {
            "maxLength": 50,
            "minLength": 2,
            "type": "string"
          },
          "location_id": {
            "type": "string"
          },
          "phone": {
            "maxLength": 20,
            "minLength": 10,
            "type": "string"
          },
          "position": {
            "maxLength": 100,
            "minLength": 2,
            "type": "string"
          },
          "specialization": {
            "maxLength": 100,
            "type": "string"
          }
        },
        "required": [
          "first_name",
          "last_name",
          "position"
        ],
        "type": "object"
      },
...
      "github_com_ialekseychuk_my-place_internal_dto.ErrorResponse": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "github_com_ialekseychuk_my-place_internal_dto.StaffResponse": {
        "properties": {
          "business_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "full_name": {
            "description": "computed field: FirstName + LastName",
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "last_name": {
            "type": "string"
          },
          "location_id": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "specialization": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "github_com_ialekseychuk_my-place_internal_dto.StaffScheduleStatsResponse": {
        "properties": {
          "average_hours_per_day": {
            "type": "number"
          },
          "period_end": {
            "type": "string"
          },
          "period_start": {
            "type": "string"
          },
          "sick_leave_days": {
            "type": "integer"
          },
          "staff_id": {
            "type": "string"
          },
          "staff_name": {
            "type": "string"
          },
          "total_overtime_hours": {
            "type": "number"
          },
          "total_shifts": {
            "type": "integer"
          },
          "total_time_off_days": {
            "type": "integer"
          },
          "total_working_hours": {
            "type": "number"
          },
          "utilization_rate": {
            "type": "number"
          },
          "vacation_days": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "github_com_ialekseychuk_my-place_internal_dto.TimeOffResponse": {
        "properties": {
          "approved_by": {
            "type": "string"
          },
          "comments": {
            "type": "string"
          },
          "end_date": {
            "type": "string"
          },
          "half_day_type": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "is_half_day": {
            "type": "boolean"
          },
          "processed_at": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "requested_at": {
            "type": "string"
          },
          "requested_by": {
            "type": "string"
          },
          "staff_id": {
            "type": "string"
          },
          "staff_name": {
            "type": "string"
          },
          "start_date": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "github_com_ialekseychuk_my-place_internal_dto.UpdateStaffRequest": {
        "properties": {
          "description": {
            "maxLength": 500,
            "type": "string"
          },
          "first_name": {
            "maxLength": 50,
            "minLength": 2,
            "type": "string"
          },
          "gender": {
            "enum": [
              "male",
              "female",
              "other",
              "prefer_not_to_say"
            ],
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "last_name": {
            "maxLength": 50,
            "minLength": 2,
            "type": "string"
          },
          "location_id": {
            "type": "string"
          },
          "phone": {
            "maxLength": 20,
            "minLength": 10,
            "type": "string"
          },
          "position": {
            "maxLength": 100,
            "minLength": 2,
            "type": "string"
          },
          "specialization": {
            "maxLength": 100,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "description": "Bearer token (JWT) for authentication. Format: Bearer \u003ctoken\u003e",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "This is a sample API with OpenAPI documentation.",
    "title": "MyPlace API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/businesses/{businessID}/schedule/stats/staff/{staffID}": {
      "get": {
        "description": "Get detailed schedule statistics for a specific staff member",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Staff ID",
            "in": "path",
            "name": "staffID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Start date (YYYY-MM-DD)",
            "in": "query",
            "name": "start_date",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "in": "query",
            "name": "end_date",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.StaffScheduleStatsResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Staff not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get staff schedule statistics",
        "tags": [
          "Schedule"
        ]
      }
    },
...
    "/api/v1/businesses/{businessID}/schedule/time-off/staff/{staffID}": {
      "get": {
        "description": "Get all time off requests for a specific staff member",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Staff ID",
            "in": "path",
            "name": "staffID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Start date filter (YYYY-MM-DD)",
            "in": "query",
            "name": "start_date",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "End date filter (YYYY-MM-DD)",
            "in": "query",
            "name": "end_date",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.TimeOffResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Staff not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get staff time off requests",
        "tags": [
          "Schedule"
        ]
      }
    },
...
    "/api/v1/businesses/{businessID}/staffs": {
      "get": {
        "description": "get staffs",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.StaffResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Validation errors"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "get staffs",
        "tags": [
          "Staff"
        ]
      },
      "post": {
        "description": "Creates a new staff for business",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.CreateStaffRequest"
              }
            }
          },
          "description": "Staff object",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.StaffResponse"
                }
              }
            },
            "description": "Created"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Validation errors"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Create a new staff",
        "tags": [
          "Staff"
        ]
      }
    },
    "/api/v1/businesses/{businessID}/staffs/{staffID}": {
      "get": {
        "description": "Get staff member by ID",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Staff ID",
            "in": "path",
            "name": "staffID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.StaffResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Staff not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get staff by ID",
        "tags": [
          "Staff"
        ]
      },
      "put": {
        "description": "Update staff member information",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Staff ID",
            "in": "path",
            "name": "staffID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.UpdateStaffRequest"
              }
            }
          },
          "description": "Staff update data",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.StaffResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Staff not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Validation errors"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Update staff",
        "tags": [
          "Staff"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}